+++ src/main/java/lach_01298/qmd/multiblock/network/DeceleratorUpdatePacket.java
@@ -5,7 +5,6 @@ import java.util.List;
 import io.netty.buffer.ByteBuf;
 import lach_01298.qmd.accelerator.Accelerator;
 import lach_01298.qmd.accelerator.tile.TileDeceleratorController;
-import lach_01298.qmd.accelerator.tile.TileRingAcceleratorController;
 import lach_01298.qmd.particle.ParticleStorageAccelerator;
 import nc.multiblock.network.MultiblockUpdatePacket;
 import nc.tile.internal.energy.EnergyStorage;
@@ -23,7 +22,7 @@ public class DeceleratorUpdatePacket extends AcceleratorUpdatePacket
 	}
 
 	public DeceleratorUpdatePacket(BlockPos pos, boolean isAcceleratorOn, long cooling, long rawHeating,
-			double maxCoolantIn, double maxCoolantOut,int maxOperatingTemp, int requiredEnergy, double efficiency, int acceleratingVoltage,
+			int maxCoolantIn, int maxCoolantOut,int maxOperatingTemp, int requiredEnergy, double efficiency, int acceleratingVoltage,
 			int RFCavityNumber, int quadrupoleNumber, double quadrupoleStrength, int dipoleNumber, double dipoleStrength,
 			int errorCode, HeatBuffer heatBuffer, EnergyStorage energyStorage, List<Tank> tanks,
 			List<ParticleStorageAccelerator> beams)
