+++ src/main/java/lach_01298/qmd/multiblock/network/LinearAcceleratorUpdatePacket.java
@@ -20,7 +20,7 @@ public class LinearAcceleratorUpdatePacket extends AcceleratorUpdatePacket
 		super();
 	}
 	
-	public LinearAcceleratorUpdatePacket(BlockPos pos, boolean isAcceleratorOn, long cooling, long rawHeating, double maxCoolantIn, double maxCoolantOut,
+	public LinearAcceleratorUpdatePacket(BlockPos pos, boolean isAcceleratorOn, long cooling, long rawHeating, int maxCoolantIn, int maxCoolantOut,
 			int maxOperatingTemp, int requiredEnergy, double efficiency, int acceleratingVoltage, int RFCavityNumber, int quadrupoleNumber,
 			double quadrupoleStrength, int dipoleNumber, double dipoleStrength, int errorCode ,HeatBuffer heatBuffer, EnergyStorage energyStorage, List<Tank> tanks,List<ParticleStorageAccelerator> beams)
 	{
